==> 01_json_scraping.ipynb <==
import json

file_path = "../../data/json_handling/spotify_playlist.json"

with open(file_path, "r") as json_file:
    data = json.load(json_file)

data

data['contents']['items'][30]['name']

for track in data['contents']['items']:
    print(track['playCount'])
    print(track['name'])
    print()
