==> 04_get_weather_data.ipynb <==
from datetime import datetime
from dateutil.relativedelta import relativedelta

import pandas as pd
import requests

url = "https://archive-api.open-meteo.com/v1/era5?latitude=52.52&longitude=13.41&start_date=2021-01-01&end_date=2021-12-31&hourly=temperature_2m"

response = requests.get(url)

response.json()

# Extract part

current_datetime = datetime.now() - relativedelta(months=8)

formatted_datetime = current_datetime.strftime("%Y-%m-%d")

url = "https://archive-api.open-meteo.com/v1/era5"


params = {"latitude": 41.85,
          "longitude": -87.65,
          "start_date": formatted_datetime,
          "end_date": formatted_datetime,
          "hourly": "temperature_2m,wind_speed_10m,precipitation,rain"}

response = requests.get(url, params=params)

weather_data = response.json()
weather_data

weather_data_filtered = {
"datetime": weather_data["hourly"]["time"],
"temperature": weather_data["hourly"]["temperature_2m"],
"wind_speed": weather_data["hourly"]["wind_speed_10m"],
"rain": weather_data["hourly"]["rain"],
"precipitation": weather_data["hourly"]["precipitation"],
}

weather_data_filtered

weather_df = pd.DataFrame(weather_data_filtered)

weather_df["datetime"] = pd.to_datetime(weather_df["datetime"])
weather_df.head()

weather_df.info()

# Transform part

weather_data_filtered = {
"datetime": weather_data["hourly"]["time"],
"temperature": weather_data["hourly"]["temperature_2m"],
"wind_speed": weather_data["hourly"]["wind_speed_10m"],
"rain": weather_data["hourly"]["rain"],
"precipitation": weather_data["hourly"]["precipitation"],
}

weather_df = pd.DataFrame(weather_data_filtered)

weather_df["datetime"] = pd.to_datetime(weather_df["datetime"])

weather_df.to_csv("weather_df_date.csv", index=False)
